consolidate decoded value representations
