+++ src/types/mod.rs
@@ -1,3 +1,4 @@
 pub mod config;
+pub mod decoded;
 pub mod shared;
 pub mod uniswap;
